+++ src/Service/MediaConvert/src/ValueObject/DynamicAudioSelector.php
@@ -10,9 +10,9 @@ use AsyncAws\MediaConvert\Enum\LanguageCode;
 /**
  * Use Dynamic audio selectors when you do not know the track layout of your source when you submit your job, but want
  * to select multiple audio tracks. When you include an audio track in your output and specify this Dynamic audio
- * selector as the Audio source, MediaConvert creates an output audio track for each dynamically selected track. Note
- * that when you include a Dynamic audio selector for two or more inputs, each input must have the same number of audio
- * tracks and audio channels.
+ * selector as the Audio source, MediaConvert creates an audio track within that output for each dynamically selected
+ * track. Note that when you include a Dynamic audio selector for two or more inputs, each input must have the same
+ * number of audio tracks and audio channels.
  */
 final class DynamicAudioSelector
 {