+++ src/Service/MediaConvert/src/ValueObject/OutputGroupSettings.php
@@ -3,6 +3,7 @@
 namespace AsyncAws\MediaConvert\ValueObject;
 
 use AsyncAws\Core\Exception\InvalidArgument;
+use AsyncAws\MediaConvert\Enum\FrameMetricType;
 use AsyncAws\MediaConvert\Enum\OutputGroupType;
 
 /**
@@ -51,6 +52,24 @@ final class OutputGroupSettings
     private $msSmoothGroupSettings;
 
     /**
+     * Optionally choose one or more per frame metric reports to generate along with your output. You can use these metrics
+     * to analyze your video output according to one or more commonly used image quality metrics. You can specify per frame
+     * metrics for output groups or for individual outputs. When you do, MediaConvert writes a CSV (Comma-Separated Values)
+     * file to your S3 output destination, named after the video, video codec, and metric type. For example:
+     * video_h264_PSNR.csv Jobs that generate per frame metrics will take longer to complete, depending on the resolution
+     * and complexity of your output. For example, some 4K jobs might take up to twice as long to complete. Note that when
+     * analyzing the video quality of your output, or when comparing the video quality of multiple different outputs, we
+     * generally also recommend a detailed visual review in a controlled environment. You can choose from the following per
+     * frame metrics: * PSNR: Peak Signal-to-Noise Ratio * SSIM: Structural Similarity Index Measure * MS_SSIM: Multi-Scale
+     * Similarity Index Measure * PSNR_HVS: Peak Signal-to-Noise Ratio, Human Visual System * VMAF: Video Multi-Method
+     * Assessment Fusion * QVBR: Quality-Defined Variable Bitrate. This option is only available when your output uses the
+     * QVBR rate control mode.
+     *
+     * @var list<FrameMetricType::*>|null
+     */
+    private $perFrameMetrics;
+
+    /**
      * Type of output group (File group, Apple HLS, DASH ISO, Microsoft Smooth Streaming, CMAF).
      *
      * @var OutputGroupType::*|null
@@ -64,6 +83,7 @@ final class OutputGroupSettings
      *   FileGroupSettings?: null|FileGroupSettings|array,
      *   HlsGroupSettings?: null|HlsGroupSettings|array,
      *   MsSmoothGroupSettings?: null|MsSmoothGroupSettings|array,
+     *   PerFrameMetrics?: null|array<FrameMetricType::*>,
      *   Type?: null|OutputGroupType::*,
      * } $input
      */
@@ -74,6 +94,7 @@ final class OutputGroupSettings
         $this->fileGroupSettings = isset($input['FileGroupSettings']) ? FileGroupSettings::create($input['FileGroupSettings']) : null;
         $this->hlsGroupSettings = isset($input['HlsGroupSettings']) ? HlsGroupSettings::create($input['HlsGroupSettings']) : null;
         $this->msSmoothGroupSettings = isset($input['MsSmoothGroupSettings']) ? MsSmoothGroupSettings::create($input['MsSmoothGroupSettings']) : null;
+        $this->perFrameMetrics = $input['PerFrameMetrics'] ?? null;
         $this->type = $input['Type'] ?? null;
     }
 
@@ -84,6 +105,7 @@ final class OutputGroupSettings
      *   FileGroupSettings?: null|FileGroupSettings|array,
      *   HlsGroupSettings?: null|HlsGroupSettings|array,
      *   MsSmoothGroupSettings?: null|MsSmoothGroupSettings|array,
+     *   PerFrameMetrics?: null|array<FrameMetricType::*>,
      *   Type?: null|OutputGroupType::*,
      * }|OutputGroupSettings $input
      */
@@ -118,6 +140,14 @@ final class OutputGroupSettings
     }
 
     /**
+     * @return list<FrameMetricType::*>
+     */
+    public function getPerFrameMetrics(): array
+    {
+        return $this->perFrameMetrics ?? [];
+    }
+
+    /**
      * @return OutputGroupType::*|null
      */
     public function getType(): ?string
@@ -145,6 +175,17 @@ final class OutputGroupSettings
         }
         if (null !== $v = $this->msSmoothGroupSettings) {
             $payload['msSmoothGroupSettings'] = $v->requestBody();
+        }
+        if (null !== $v = $this->perFrameMetrics) {
+            $index = -1;
+            $payload['perFrameMetrics'] = [];
+            foreach ($v as $listValue) {
+                ++$index;
+                if (!FrameMetricType::exists($listValue)) {
+                    throw new InvalidArgument(\sprintf('Invalid parameter "perFrameMetrics" for "%s". The value "%s" is not a valid "FrameMetricType".', __CLASS__, $listValue));
+                }
+                $payload['perFrameMetrics'][$index] = $listValue;
+            }
         }
         if (null !== $v = $this->type) {
             if (!OutputGroupType::exists($v)) {