+++ src/Service/MediaConvert/src/ValueObject/ProresSettings.php
@@ -3,6 +3,7 @@
 namespace AsyncAws\MediaConvert\ValueObject;
 
 use AsyncAws\Core\Exception\InvalidArgument;
+use AsyncAws\MediaConvert\Enum\FrameMetricType;
 use AsyncAws\MediaConvert\Enum\ProresChromaSampling;
 use AsyncAws\MediaConvert\Enum\ProresCodecProfile;
 use AsyncAws\MediaConvert\Enum\ProresFramerateControl;
@@ -128,6 +129,24 @@ final class ProresSettings
     private $parNumerator;
 
     /**
+     * Optionally choose one or more per frame metric reports to generate along with your output. You can use these metrics
+     * to analyze your video output according to one or more commonly used image quality metrics. You can specify per frame
+     * metrics for output groups or for individual outputs. When you do, MediaConvert writes a CSV (Comma-Separated Values)
+     * file to your S3 output destination, named after the video, video codec, and metric type. For example:
+     * video_h264_PSNR.csv Jobs that generate per frame metrics will take longer to complete, depending on the resolution
+     * and complexity of your output. For example, some 4K jobs might take up to twice as long to complete. Note that when
+     * analyzing the video quality of your output, or when comparing the video quality of multiple different outputs, we
+     * generally also recommend a detailed visual review in a controlled environment. You can choose from the following per
+     * frame metrics: * PSNR: Peak Signal-to-Noise Ratio * SSIM: Structural Similarity Index Measure * MS_SSIM: Multi-Scale
+     * Similarity Index Measure * PSNR_HVS: Peak Signal-to-Noise Ratio, Human Visual System * VMAF: Video Multi-Method
+     * Assessment Fusion * QVBR: Quality-Defined Variable Bitrate. This option is only available when your output uses the
+     * QVBR rate control mode.
+     *
+     * @var list<FrameMetricType::*>|null
+     */
+    private $perFrameMetrics;
+
+    /**
      * Use this setting for interlaced outputs, when your output frame rate is half of your input frame rate. In this
      * situation, choose Optimized interlacing to create a better quality interlaced output. In this case, each progressive
      * frame from the input corresponds to an interlaced field in the output. Keep the default value, Basic interlacing, for
@@ -173,6 +192,7 @@ final class ProresSettings
      *   ParControl?: null|ProresParControl::*,
      *   ParDenominator?: null|int,
      *   ParNumerator?: null|int,
+     *   PerFrameMetrics?: null|array<FrameMetricType::*>,
      *   ScanTypeConversionMode?: null|ProresScanTypeConversionMode::*,
      *   SlowPal?: null|ProresSlowPal::*,
      *   Telecine?: null|ProresTelecine::*,
@@ -190,6 +210,7 @@ final class ProresSettings
         $this->parControl = $input['ParControl'] ?? null;
         $this->parDenominator = $input['ParDenominator'] ?? null;
         $this->parNumerator = $input['ParNumerator'] ?? null;
+        $this->perFrameMetrics = $input['PerFrameMetrics'] ?? null;
         $this->scanTypeConversionMode = $input['ScanTypeConversionMode'] ?? null;
         $this->slowPal = $input['SlowPal'] ?? null;
         $this->telecine = $input['Telecine'] ?? null;
@@ -207,6 +228,7 @@ final class ProresSettings
      *   ParControl?: null|ProresParControl::*,
      *   ParDenominator?: null|int,
      *   ParNumerator?: null|int,
+     *   PerFrameMetrics?: null|array<FrameMetricType::*>,
      *   ScanTypeConversionMode?: null|ProresScanTypeConversionMode::*,
      *   SlowPal?: null|ProresSlowPal::*,
      *   Telecine?: null|ProresTelecine::*,
@@ -286,6 +308,14 @@ final class ProresSettings
     }
 
     /**
+     * @return list<FrameMetricType::*>
+     */
+    public function getPerFrameMetrics(): array
+    {
+        return $this->perFrameMetrics ?? [];
+    }
+
+    /**
      * @return ProresScanTypeConversionMode::*|null
      */
     public function getScanTypeConversionMode(): ?string
@@ -362,6 +392,17 @@ final class ProresSettings
         }
         if (null !== $v = $this->parNumerator) {
             $payload['parNumerator'] = $v;
+        }
+        if (null !== $v = $this->perFrameMetrics) {
+            $index = -1;
+            $payload['perFrameMetrics'] = [];
+            foreach ($v as $listValue) {
+                ++$index;
+                if (!FrameMetricType::exists($listValue)) {
+                    throw new InvalidArgument(\sprintf('Invalid parameter "perFrameMetrics" for "%s". The value "%s" is not a valid "FrameMetricType".', __CLASS__, $listValue));
+                }
+                $payload['perFrameMetrics'][$index] = $listValue;
+            }
         }
         if (null !== $v = $this->scanTypeConversionMode) {
             if (!ProresScanTypeConversionMode::exists($v)) {