+++ src/Service/MediaConvert/src/ValueObject/XavcSettings.php
@@ -3,6 +3,7 @@
 namespace AsyncAws\MediaConvert\ValueObject;
 
 use AsyncAws\Core\Exception\InvalidArgument;
+use AsyncAws\MediaConvert\Enum\FrameMetricType;
 use AsyncAws\MediaConvert\Enum\XavcAdaptiveQuantization;
 use AsyncAws\MediaConvert\Enum\XavcEntropyEncoding;
 use AsyncAws\MediaConvert\Enum\XavcFramerateControl;
@@ -83,6 +84,24 @@ final class XavcSettings
      * @var int|null
      */
     private $framerateNumerator;
+
+    /**
+     * Optionally choose one or more per frame metric reports to generate along with your output. You can use these metrics
+     * to analyze your video output according to one or more commonly used image quality metrics. You can specify per frame
+     * metrics for output groups or for individual outputs. When you do, MediaConvert writes a CSV (Comma-Separated Values)
+     * file to your S3 output destination, named after the video, video codec, and metric type. For example:
+     * video_h264_PSNR.csv Jobs that generate per frame metrics will take longer to complete, depending on the resolution
+     * and complexity of your output. For example, some 4K jobs might take up to twice as long to complete. Note that when
+     * analyzing the video quality of your output, or when comparing the video quality of multiple different outputs, we
+     * generally also recommend a detailed visual review in a controlled environment. You can choose from the following per
+     * frame metrics: * PSNR: Peak Signal-to-Noise Ratio * SSIM: Structural Similarity Index Measure * MS_SSIM: Multi-Scale
+     * Similarity Index Measure * PSNR_HVS: Peak Signal-to-Noise Ratio, Human Visual System * VMAF: Video Multi-Method
+     * Assessment Fusion * QVBR: Quality-Defined Variable Bitrate. This option is only available when your output uses the
+     * QVBR rate control mode.
+     *
+     * @var list<FrameMetricType::*>|null
+     */
+    private $perFrameMetrics;
 
     /**
      * Specify the XAVC profile for this output. For more information, see the Sony documentation at
@@ -194,6 +213,7 @@ final class XavcSettings
      *   FramerateConversionAlgorithm?: null|XavcFramerateConversionAlgorithm::*,
      *   FramerateDenominator?: null|int,
      *   FramerateNumerator?: null|int,
+     *   PerFrameMetrics?: null|array<FrameMetricType::*>,
      *   Profile?: null|XavcProfile::*,
      *   SlowPal?: null|XavcSlowPal::*,
      *   Softness?: null|int,
@@ -214,6 +234,7 @@ final class XavcSettings
         $this->framerateConversionAlgorithm = $input['FramerateConversionAlgorithm'] ?? null;
         $this->framerateDenominator = $input['FramerateDenominator'] ?? null;
         $this->framerateNumerator = $input['FramerateNumerator'] ?? null;
+        $this->perFrameMetrics = $input['PerFrameMetrics'] ?? null;
         $this->profile = $input['Profile'] ?? null;
         $this->slowPal = $input['SlowPal'] ?? null;
         $this->softness = $input['Softness'] ?? null;
@@ -234,6 +255,7 @@ final class XavcSettings
      *   FramerateConversionAlgorithm?: null|XavcFramerateConversionAlgorithm::*,
      *   FramerateDenominator?: null|int,
      *   FramerateNumerator?: null|int,
+     *   PerFrameMetrics?: null|array<FrameMetricType::*>,
      *   Profile?: null|XavcProfile::*,
      *   SlowPal?: null|XavcSlowPal::*,
      *   Softness?: null|int,
@@ -291,6 +313,14 @@ final class XavcSettings
     public function getFramerateNumerator(): ?int
     {
         return $this->framerateNumerator;
+    }
+
+    /**
+     * @return list<FrameMetricType::*>
+     */
+    public function getPerFrameMetrics(): array
+    {
+        return $this->perFrameMetrics ?? [];
     }
 
     /**
@@ -390,6 +420,17 @@ final class XavcSettings
         }
         if (null !== $v = $this->framerateNumerator) {
             $payload['framerateNumerator'] = $v;
+        }
+        if (null !== $v = $this->perFrameMetrics) {
+            $index = -1;
+            $payload['perFrameMetrics'] = [];
+            foreach ($v as $listValue) {
+                ++$index;
+                if (!FrameMetricType::exists($listValue)) {
+                    throw new InvalidArgument(\sprintf('Invalid parameter "perFrameMetrics" for "%s". The value "%s" is not a valid "FrameMetricType".', __CLASS__, $listValue));
+                }
+                $payload['perFrameMetrics'][$index] = $listValue;
+            }
         }
         if (null !== $v = $this->profile) {
             if (!XavcProfile::exists($v)) {